pyproject.toml
[tool.poetry]
name = "service_download"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
common = "0.0.19"
logger = "0.1.0"
anyio = "3.4.0"
fastapi = "0.62.0"
fastapi-utils = "0.2.1"
uvicorn = "0.12.3"
gunicorn = "20.0.4"
uvloop = "0.14.0"
httptools = "0.1.1"
PyJWT = "1.4.2"
python-multipart = "0.0.5"
python-json-logger = "0.1.11"
minio = "7.0.3"
xmltodict = "0.12.0"
pydantic = "1.8.2"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "0.28b1"
opentelemetry-instrumentation-fastapi = "0.28b1"
opentelemetry-instrumentation-sqlalchemy = "0.28b1"
opentelemetry-instrumentation-requests = "0.28b1"
opentelemetry-instrumentation-asyncpg = "0.28b1"
opentelemetry-instrumentation-httpx = "0.28b1"
sqlalchemy = {extras = ["asyncio"], version = "^1.4.32"}
psycopg2-binary = "2.9.2"
asyncpg = "0.25.0"
httpx = "^0.22.0"
aioredis = "^2.0.1"
aiofiles = "^0.8.0"
greenlet = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
Faker = "11.3.0"
pytest-asyncio = "^0.18.1"
pytest-env = "0.6.2"
pytest-cov = "^3.0.0"
pytest-httpx = "^0.19.0"
flake8 = "^4.0.1"
async-asgi-testclient = "^1.4.9"
testcontainers = "^3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pilot"
url = "https://git.indocresearch.org/api/v4/groups/pilot/-/packages/pypi/simple"
default = true
